move progressbar to _vendor
