+++ conda/_vendor/progressbar/compat.py
