+++ conda/_vendor/progressbar/widgets.py
